@@ -1,11 +1,12 @@
 class PicksController < ApplicationController
   def create
-    @pick = Pick.create!(params[:pick])
+    @pick = Pick.create!(params.require(:pick).permit!)
+
     redirect_to_list_url
   end
 
   def destroy
-    @pick = Pick.find(params[:id])
+    @pick = Pick.find(params.require(:id).permit)
     @pick.delete
     redirect_to_list_url
   end
